+++ main/serializers/__init__.py
@@ -1,4 +1,4 @@
-from main.serializers.serializer_address_search import *
+from main.serializers.serializer_address_scan import *
 from main.serializers.serializer_balance import *
 from main.serializers.serializer_blockheight import *
 from main.serializers.serializer_token import *
