+++ main/serializers/serializer_address_scan.py
@@ -1,5 +1,5 @@
 from rest_framework import serializers
-from main.utils.address_search import sort_address_sets
+from main.utils.address_scan import sort_address_sets
 
 
 class AddressSetSerializer(serializers.Serializer):
@@ -12,7 +12,7 @@ class WalletAddressSetSerializer(serializers.Serializer):
     addresses = AddressSetSerializer()
 
 
-class WalletAddressSearchSerializer(serializers.Serializer):
+class WalletAddressScanSerializer(serializers.Serializer):
     wallet_hash = serializers.CharField(max_length=200)
     project_id = serializers.CharField(max_length=200, required=False, allow_blank=True)
     address_sets = WalletAddressSetSerializer(many=True)
@@ -22,7 +22,7 @@ class WalletAddressSearchSerializer(serializers.Serializer):
         return sort_address_sets(address_sets)
 
 
-class WalletAddressSearchResponseSerializer(serializers.Serializer):
+class WalletAddressScanResponseSerializer(serializers.Serializer):
     address_set = WalletAddressSetSerializer()
     success = serializers.BooleanField()
     error = serializers.CharField(required=False, allow_blank=True)
