+++ main/views/__init__.py
@@ -1,4 +1,4 @@
-from main.views.view_address_search import *
+from main.views.view_address_scan import *
 from main.views.view_telegram import *
 from main.views.view_subscribe import *
 from main.views.view_test import *
